compare python and libtorch classifier; add RUN.log

+++ include/gpd/net/libtorch_classifier.h
@@ -47,7 +47,7 @@ namespace gpd {
              * \param image_list the list of grasp images
              * \return the classified grasp candidates
              */
-            std::vector<float> classifyImages(
+            std::vector<double> classifyImages(
                     const std::vector<std::unique_ptr<cv::Mat>> &image_list);
 
             /**
@@ -55,7 +55,7 @@ namespace gpd {
              * \param point_list the points in the hand closed area.
              * \return the classified grasp candidates
              */
-            std::vector<float> classifyPoints(
+            std::vector<double> classifyPoints(
                     const std::vector<std::unique_ptr<Eigen::Matrix3Xd>> &point_groups);
 
             /**
@@ -63,7 +63,7 @@ namespace gpd {
              * \param point_list the points in the hand closed area.
              * \return the classified grasp candidates
              */
-            std::vector<float> classifyPointsBatch(
+            std::vector<double> classifyPointsBatch(
                     const std::vector<std::unique_ptr<Eigen::Matrix3Xd>> &point_groups);
 
             int getBatchSize() const { return batch_size_; }
